Revised rental model for the slices functions
