+++ src/models/requests/AddRentalRequestModel.ts
@@ -1,0 +1,6 @@
+export interface AddRentalRequestModel {
+	startDate : Date| null;
+	endDate : Date| null;
+	carId: number;
+	userId: number;
+}
