+++ src/models/requests/RentalIdRequestModel.ts
@@ -1,0 +1,6 @@
+export interface RentalIdRequestModel {
+	startDate : string;
+	endDate : string;
+	carId: number;
+	userId: number;
+}
