+++ src/models/response/GetRentalIdModel.ts
@@ -1,0 +1,3 @@
+export interface GetRentalIdModel {
+    id: number,
+}
