+++ src/models/response/RentalResponseModel.ts
@@ -1,0 +1,12 @@
+import { CarModel } from "../response/CarModel";
+
+export interface RentalResponseModel {
+  id: number;
+  startDate: Date;
+  endDate: Date;
+  returnDate?: Date | null;
+  startKilometer: number;
+  totalPrice: number;
+  car: CarModel;
+  userId: number;
+}
